(svn r3661) -Fix [PPC computers] quick dirty fix for failure to start the game on BE CPUs 	3626 broke the game on Big endian CPUs and this is a quick dirty fix 	so the nightly builds for OSX and MorphOS will work again 	the game will still fail on 64 bit BE CPUs 	note: the game runs in 32 bit mode on G5, so it will work on G5 	we need to make a better fix for this, but we also need the nightly builds to work

+++ saveload.h
@@ -136,7 +136,7 @@ enum SaveLoadTypes {
 };
 
 typedef byte SaveLoadType;
-typedef uint16 OffSetType;
+typedef uint32 OffSetType;
 
 /** SaveLoad type struct. Do NOT use this directly but use the SLE_ macros defined just below! */
 typedef struct SaveLoad {
